refactor: rename route and its related bits
I think I called the page choose ages when it made more sense. But no longer!
