+++ server/app.js
@@ -8,7 +8,7 @@ import { controller as catchErrors } from './pages/error-catch-all.js';
 import { controller as notFound } from './pages/error-not-found.js';
 import { controller as home } from './pages/home.js';
 import { controller as games } from './pages/games.js';
-import { controller as choose } from './pages/choose.js';
+import { controller as share } from './pages/share.js';
 import { controller as turn } from './pages/turn.js';
 import game from './pages/game/controller.js';
 import { read as apiRead, update as apiUpdate } from './pages/api-proxy.js';
@@ -26,7 +26,7 @@ app.post(
 	express.urlencoded({ extended: false }),
 	catchRejections(games)
 );
-app.get('/games/:gameId/choose', catchRejections(choose));
+app.get('/games/:gameId/share', catchRejections(share));
 app.get('/games/:gameId', catchRejections(game));
 app.post(
 	'/games/:gameId/turn',
